działający kod FIFO
